comments, fixed some IO stuff, trashing chunks

+++ helpers/OShelper.cpp
@@ -6,12 +6,6 @@
 #include <string.h>
 
 
-/*!
- * \brief OSHelper::checkFile
- * \param path - path to the file
- * \return 0 in case of success
- * checks, whether the file is known video file
- */
 int32_t OSHelper::checkFile(std::string &path) {
     struct stat info;
 
@@ -19,11 +13,13 @@ int32_t OSHelper::checkFile(std::string &path) {
         reportError("An error occured while controlling the file: " + path);
         return (-1);
     }
+    // part after the last dot
     std::string extension = getExtension(path);
     if (extension.empty()) {
         reportError("Invalid file extension.");
         return (-1);
     }
+    // checks whether supported
     std::vector<std::string> video_ext = DATA->getKnownFormats();
     if (find(video_ext.begin(), video_ext.end(), extension) == video_ext.end()) {
         reportError("Unknown file extension");
@@ -32,14 +28,6 @@ int32_t OSHelper::checkFile(std::string &path) {
     return (0);
 }
 
-/*!
- * \brief OSHelper::rmrDir remove directory
- * \param dir path to the directory
- * \param recursive whether delete content recursively
- * \return  0 in case of success
- * removes content of the directory and the directory itself
- * recursive deletion need to be specified
- */
 int32_t OSHelper::rmrDir(std::string dir, bool recursive) {
   DIR *d, *t;
   struct dirent *entry;
@@ -47,80 +35,85 @@ int32_t OSHelper::rmrDir(std::string dir, bool recursive) {
   d = opendir(dir.c_str());
   if (d == NULL)
     return (-1);
+  // loops through all entries in the directory
   while ((entry = readdir(d))) {
+      // ignoring special cases
     if ((strcmp(entry->d_name, ".") == 0) ||
       (strcmp(entry->d_name, "..") == 0))
       continue;
     sprintf(abs_fn, "%s/%s", dir.c_str(), entry->d_name);
+
+    // if the entry is a directory
     if ((t = opendir(abs_fn))) {
       closedir(t);
-      if(recursive)
+      // if deletion should dive recursively
+      if (recursive) {
         rmrDir(abs_fn, true);
-      else
-         return (-1);
+      } else {
+         return -1;
+      }
+
+     // entry is not a directory
     } else {
-      if (unlink(abs_fn) == -1)
-          return (-1);
+        // remove the file
+      if (unlink(abs_fn) == -1) {
+          return -1;
+      }
     }
   }
   closedir(d);
+  // finally removes the directory itself
   rmdir(dir.c_str());
   return (0);
 }
 
-/*!
- * \brief OSHelper::rmFile removes file
- * \param fp path to the file
- * \return 0 on success
- */
 int32_t OSHelper::rmFile(std::string fp) {
+    // file does not exist - nothing to remove
+    if (!OSHelper::isFileOk(fp)) {
+        return 0;
+    }
     if (unlink(fp.c_str()) == -1) {
-        reportDebug(fp + ": Failed to remove.", 1);
         return -1;
     }
     return 0;
 }
 
-/*!
- * \brief OSHelper::prepareDir ensures the given directory exists
- * \param location location to the desired directory
- * \param destroy whether delete contents in case of existence
- * \return 0 on success
- * creates the desired directory, and others in path if needed
- * existing directory is treated according to the destroy parameter
- */
 int32_t OSHelper::prepareDir(std::string location, bool destroy) {
     std::string directory, current;
+    // the WD should definitely exist already
     location = location.substr(
-                DATA->config.getStringValue("WD").size() + 1, location.length());
+                DATA->config.getStringValue("WD").size() + 1,
+                location.length());
     uint32_t pos;
+    // holds the current position
     current = DATA->config.getStringValue("WD");
+    // further directories on the path
     while ((pos = location.find('/')) != std::string::npos) {
+        // extract next directory
         directory = location.substr(0, pos);
+        // shorten the location - holds the rest og the path
         if ( location.find('/') != std::string::npos) {
             location = location.substr(pos + 1, location.length());
         }
+        // update current
         current += "/" + directory;
-        //todo remove rest?
+        // creates the directory
         if (OSHelper::mkDir(current, destroy) == -1) {
             return -1;
         }
     }
+    // eventually the location contains the last directory
     current += "/" + location;
     return OSHelper::mkDir(current, destroy);
 }
 
-/*!
- * \brief OSHelper::mkDir creates directory
- * \param location path to the directory
- * \param destroy whether remove existing
- * \return 0 on success
- */
-
 int32_t OSHelper::mkDir(std::string location, bool destroy) {
     if (mkdir(location.c_str(), 0700) == -1) {
         switch (errno) {
+        // the directory existed already
         case EEXIST:
+            // if destroy, the existing direcotry is removed recursively
+            // end with error otherwise
             if (destroy) {
                 if (rmrDir(location.c_str(), false) == -1)
                     return -1;
@@ -136,17 +129,6 @@ int32_t OSHelper::mkDir(std::string location, bool destroy) {
     return 0;
 }
 
-/*!
- * \brief OSHelper::runExternal spawns external command
- * \param stdo string to save stdout
- * \param stde string to save stderr
- * \param cmd name of the command to spawn
- * \param numargs number of parameters
- * \return 0 on success
- * spawns the xternal command, parameters are given optionally
- * uses fork & exec, waits for the result_type
- * stderr and stdout are saved in the corresponding strings
- */
 int32_t OSHelper::runExternal(std::string &stdo, std::string &stde, char *cmd, int32_t numargs, ...) {
     pid_t pid;
     int32_t pd_o[2], pd_e[2], j;
@@ -154,12 +136,16 @@ int32_t OSHelper::runExternal(std::string &stdo, std::string &stde, char *cmd, i
     std::string whole_command(cmd);
     char buf_o[bufsize], buf_e[bufsize];
     char *bo = buf_o, *be = buf_e;
+    // creates pipes
     pipe(pd_o);
     pipe(pd_e);
+    // forks
     switch (pid = fork()) {
+    // child process
     case 0: {
         va_list arg_ptr;
         va_start (arg_ptr, numargs);
+        // obtain parameters
         char *args[numargs + 3];
         // TODO: safety!
         for(j = 0 ; j < numargs; ++j) {
@@ -179,12 +165,14 @@ int32_t OSHelper::runExternal(std::string &stdo, std::string &stde, char *cmd, i
         dup(pd_e[1]);
         char command[BUF_LENGTH];
         strcpy(command, cmd);
+        // execute the desired programme
         if ((execvp(command, args)) == -1) {
             reportError("Error while spawning external command.");
-            return (-1);
+            return -1;
         }
         break;
         }
+        // error while forking
     case -1:
         close(pd_o[0]);
         close(pd_o[1]);
@@ -192,12 +180,16 @@ int32_t OSHelper::runExternal(std::string &stdo, std::string &stde, char *cmd, i
         close(pd_e[1]);
         return (-1);
         break;
+        // parent
     default:
         close(pd_o[1]);
         close(pd_e[1]);
         stdo = stde = "";
         int32_t st;
+        // wait for the child to end
         wait(&st);
+        // reads the output from the pipe to the buffer
+        // TODO: should not go output to the file?
         while(read(pd_o[0], bo, 1) == 1){
             bo++;
             //always positive anyway
@@ -207,6 +199,7 @@ int32_t OSHelper::runExternal(std::string &stdo, std::string &stde, char *cmd, i
                 bo = buf_o;
             }
         }
+        // same with stderr
         while(read(pd_e[0], be, 1) == 1){
             be++;
             if((unsigned)(be - buf_e) > bufsize){
@@ -224,58 +217,48 @@ int32_t OSHelper::runExternal(std::string &stdo, std::string &stde, char *cmd, i
     return 0;
 }
 
-/*!
- * \brief OSHelper::isFileOk checks, whether the file exists and is nonempty
- * \param fp
- * \return true if the file is ok
- */
 bool OSHelper::isFileOk(const std::string &fp) {
     struct stat st;
+    // tries to stat the file
     if (stat(fp.c_str(), &st) == -1) {
         return false;
     }
     return true;
 }
 
-/*!
- * \brief OSHelper::getFileSize
- * \param file path to the file
- * \return size of the file in bytes
- */
 int64_t OSHelper::getFileSize(const std::string &file) {
     struct stat64 finfo;
+    // TODO: really 64?
+    // negative RV indicates error
     if (lstat64(file.c_str(), &finfo) == -1)
-        return (-1);
+        return -1;
     return finfo.st_size;
 }
 
-/*!
- * \brief OSHelper::getExtension extracts file extension
- * \param str name of the file
- * \return file extension
- */
 std::string OSHelper::getExtension(const std::string &str) {
     size_t pos = str.rfind('.');
     std::string extension("");
-    if (pos != std::string::npos)
+    // extracts the part behind last period
+    if (pos != std::string::npos) {
         extension = str.substr(pos + 1);
+    }
     return extension;
 }
 
-/*!
- * \brief OSHelper::getBasename extracts the file's basename
- * \param str path to the files
- * \return basename of the file, without extension
- */
 std::string OSHelper::getBasename(const std::string &str) {
     size_t pos = str.rfind('/');
     std::string basename("");
-    if (pos == std::string::npos)
-        return std::string("");
-    basename = str.substr(pos + 1);
+    // the file itself is basename
+    if (pos == std::string::npos) {
+        basename = str;
+    } else {
+        basename = str.substr(pos + 1);
+    }
     pos = basename.rfind('.');
-    if (pos == std::string::npos)
+    // strips the extension, if any
+    if (pos == std::string::npos) {
         return basename;
+    }
     basename = basename.substr(0, pos);
     return basename;
 }
